BAEL-2682: Move java-immutable-set article to core-java-collections-set module
