+++ core-java-modules/core-java-collections-set/src/main/java/com/baeldung/set/UnmodifiableSet.java
@@ -1,4 +1,4 @@
-package com.baeldung.java9.set;
+package com.baeldung.set;
 
 import com.google.common.collect.ImmutableSet;
