+++ core-java-modules/core-java-collections-set/src/test/java/com/baeldung/set/SetExamplesUnitTest.java
@@ -1,9 +1,10 @@
-package com.baeldung.java9;
+package com.baeldung.set;
+
+import org.junit.Test;
 
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
-import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
